# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Balance API",
        "description": "Тестовое API. ",
        "version": "1.0.0"
    },
    "paths": {
        "/api/balance/{user_id}": {
            "get": {
                "tags": [
                    "Balance"
                ],
                "summary": "Получить баланс пользователя",
                "operationId": "e9ef716f5322270cf23d87d2ef299f86",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "UUID пользователя",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Текущий баланс",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        },
                                        "amount": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Баланс не найден"
                    }
                }
            }
        },
        "/api/deposit": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Пополнение счёта",
                "operationId": "a79cd46c0bb6c12a2ae0ebd5f3c71067",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Пополнение через терминал"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Депозит выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Несовпадение X-User-ID и user_id"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/withdraw": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Списание средств",
                "operationId": "08f60ad504c887ba989dfcc0a181476d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "amount"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Снятие наличных"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Вывод выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Несовпадение X-User-ID и user_id"
                    },
                    "400": {
                        "description": "Недостаточно средств"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/transfer": {
            "post": {
                "tags": [
                    "Transactions"
                ],
                "summary": "Перевод между пользователями",
                "operationId": "f9fcb9875722ec438b4cbace050c114e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "from_user_id",
                                    "to_user_id",
                                    "amount"
                                ],
                                "properties": {
                                    "from_user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "to_user_id": {
                                        "type": "string",
                                        "format": "uuid"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Перевод за обед"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Перевод выполнен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "transaction_id": {
                                            "type": "string",
                                            "format": "uuid"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Несовпадение X-User-ID и from_user_id"
                    },
                    "400": {
                        "description": "Недостаточно средств"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить список всех пользователей",
                "operationId": "566aed8efb436ca506e0ef51a032ecd9",
                "responses": {
                    "200": {
                        "description": "Список пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "e7b8c2f0-1234-4a9f-8a2b-abc123def456"
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Name"
                                            },
                                            "email": {
                                                "type": "string",
                                                "example": "name@example.com"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создать нового пользователя",
                "operationId": "a794d2c5818707a660863ff972b607ac",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Name"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "name@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "e7b8c2f0-1234-4a9f-8a2b-abc123def456"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Email уже занят",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "User with this email already exists"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Transactions",
            "description": "Операции с балансом: пополнение, вывод, перевод"
        },
        {
            "name": "Balance",
            "description": "Balance"
        },
        {
            "name": "Users",
            "description": "Users"
        }
    ]
}